+++ src/models/components/event_outcome.py
@@ -20,7 +20,7 @@ class EventOutcome(Enum):
         return self in {EventOutcome.SUCCESS}
 
     @staticmethod
-    def decide_outcome(timestamp: float, source_obj: GameObj, spell: Spell, target_obj: GameObj, skip_source_validation: bool) -> 'EventOutcome':
+    def decide_outcome(timestamp: int, source_obj: GameObj, spell: Spell, target_obj: GameObj, skip_source_validation: bool) -> 'EventOutcome':
         # Validate source
         if not skip_source_validation:
             if not source_obj.status.is_valid_source:
@@ -43,7 +43,7 @@ class EventOutcome(Enum):
         return source_obj.pos.has_target_within_range(target_obj.pos, spell.range_limit)
 
     @staticmethod
-    def _gcd_is_available(timestamp: float, source_obj: GameObj, spell: Spell) -> bool:
+    def _gcd_is_available(timestamp: int, source_obj: GameObj, spell: Spell) -> bool:
         if not spell.flags & Behavior.TRIGGER_GCD:
             return True
         return source_obj.get_gcd_progress(timestamp) >= 1.0
